# packages/api/swagger/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "signUp",
        "summary": "Register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "signIn",
        "summary": "Log In",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "changePassword",
        "summary": "Change password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyUserDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/api-keys/{id}": {
      "delete": {
        "operationId": "deleteApiKey",
        "summary": "Delete API Keys",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/generate-apikey": {
      "post": {
        "operationId": "generateApiKey",
        "summary": "Create API Key",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "operationId": "refreshAccessToken",
        "summary": "Refresh Access Token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/ticketing/users": {
      "get": {
        "operationId": "getUsers",
        "summary": "List a batch of Users",
        "parameters": [
          {
            "name": "x-connection-token",
            "required": true,
            "in": "header",
            "description": "The connection token",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "remote_data",
            "required": false,
            "in": "query",
            "description": "Set to true to include data from the original software.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Set to get the number of records.",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Set to get the number of records after this cursor.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UnifiedUserOutput"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "ticketing/users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/crm/users": {
      "get": {
        "operationId": "getUsers",
        "summary": "List a batch of Users",
        "parameters": [
          {
            "name": "x-connection-token",
            "required": true,
            "in": "header",
            "description": "The connection token",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "remote_data",
            "required": false,
            "in": "query",
            "description": "Set to true to include data from the original software.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Set to get the number of records.",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Set to get the number of records after this cursor.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UnifiedUserOutput"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "crm/users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/syncs/status/{vertical}": {
      "get": {
        "operationId": "getSyncStatus",
        "summary": "Retrieve sync status of a certain vertical",
        "parameters": [
          {
            "name": "vertical",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "syncs"
        ]
      }
    },
    "/syncs/resyncs/{user_id}/{vertical}": {
      "get": {
        "operationId": "resync",
        "summary": "Resync common objects across a vertical",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vertical",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "syncs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Unified Panora API",
    "description": "The Panora API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "strategy": {
            "type": "string"
          },
          "password_hash": {
            "type": "string",
            "minLength": 9
          },
          "id_organisation": {
            "type": "string"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "email",
          "strategy"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "id_user": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password_hash": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password_hash"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "old_password_hash": {
            "type": "string"
          },
          "new_password_hash": {
            "type": "string",
            "minLength": 9
          }
        },
        "required": [
          "email",
          "old_password_hash",
          "new_password_hash"
        ]
      },
      "VerifyUserDto": {
        "type": "object",
        "properties": {
          "id_user": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          }
        },
        "required": [
          "id_user",
          "email",
          "first_name",
          "last_name"
        ]
      },
      "ApiKeyDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "keyName": {
            "type": "string"
          }
        },
        "required": [
          "projectId",
          "userId",
          "keyName"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          }
        },
        "required": [
          "projectId"
        ]
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "statusCode": {
            "type": "number"
          }
        },
        "required": [
          "statusCode"
        ]
      },
...
      "UnifiedUserOutput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the user"
          },
          "email": {
            "type": "string",
            "description": "The email of the user"
          },
          "field_mappings": {
            "type": "object",
            "properties": {}
          },
          "id": {
            "type": "string",
            "description": "The uuid of the user"
          },
          "remote_id": {
            "type": "string",
            "description": "The id of the user in the context of the Crm 3rd Party"
          },
          "remote_data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "name",
          "email",
          "field_mappings",
          "remote_data"
        ]
      },
...
    }
  }
}